IMPALA-9627 (followup): Use -q for bootstrap_toolchain.py's wget calls
The original fix for IMPALA-9627 removed the local check_output()
function for bin/bootstrap_toolchain.py and replaced it with
subprocess.check_call(). The local check_output() function was
preventing output from the wget call from going to the commandline.
The subprocess.check_call() doesn't do this. Since we don't want
this output anyway, this modifies the wget call to add "-q" to
avoid the extra output.

Testing:
 - Verified the extra output was gone

Change-Id: If5b81f502a73551d269a0379ed0106ba4d3c8363
Reviewed-on: http://gerrit.cloudera.org:8080/19814
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>