IMPALA-12059: Deduplicate and filter warnings from Thrift compilation
This wraps Thrift compilation commands with the
bin/thrift-quiet-wrapper.sh shell script. This script runs the
Thrift compilation command, collecting the output. If the
command fails, it dumps the output as-is. If the command succeeds,
it deduplicates the warnings and filters out a couple warnings
that we have no intention of fixing.

Testing:
 - Ran "make gen-deps" on a clean build and verified that
   the Thrift warnings are deduplicated and filtered
 - Modified a Thrift file and verified errors are printed as-is.

Change-Id: I7b912ac3d57d1a51e957889b5798dc05d156a3d0
Reviewed-on: http://gerrit.cloudera.org:8080/19818
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>