Install libreadline-dev for toolchain builds on ARM hosts
This is required because ARM builds of the toolchain binaries are not
(yet) present in s3://native-toolchain, so ARM builds have to build the
toolchain locally, before being able to build Impala.

The toolchain's Python build failed before this patch, because it missed
the libreadline-dev package, which is needed for Python's readline support.
This patch adds it and its libncurses dependency.

Change-Id: I1bf6193027d691d3ded727cb59424c5dc9963ea9
Reviewed-on: http://gerrit.cloudera.org:8080/19835
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>