IMPALA-12113: Fix elapsed time calculation in query timeline
This patch changes the elapsed time at the bottom of the query timeline
to show the end of each interval. This was the original intention to use
the end time and right-align these values so that the full elapsed time
of the query is shown on the last interval. The query start time is
always implicitly 0 in the timeline.

Testing: Manually tested query in WebUI

Change-Id: I3ed33f13fbe1bf3e12e149db2e72ddbac0178643
Reviewed-on: http://gerrit.cloudera.org:8080/19834
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>