IMPALA-12121: Add non empty check for UDF location in getLastModifiedTime method
getLastModifiedTime() of Function.java assumes that if the UDF is not a BUILTIN
and if the location is not null, it is a non-empty string. However, we may have
UDFs with an empty location if they are in the install location and loaded at
init time.

This patch introduces an isBuiltinOrJava() method that checks if the function is
BUILTIN, or a JAVA function with blank location.

Change-Id: Ia6035d1a21e7222f0e95f984a0f4022ba2e89d9f
Reviewed-on: http://gerrit.cloudera.org:8080/19852
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>