IMPALA-12129: Handle running queries in query timeline
This patch fixes handling of running queries in the query timeline. An
additional check was added to skip JSON sections that are not available
while the query is running. The page refresh was changed to wait until
End Time is set since Query State can be set to FINISHED for running
queries.

Testing: Manually tested query in WebUI

Change-Id: I5a426008addc49183af1b317164d5b02d5f8ceec
Reviewed-on: http://gerrit.cloudera.org:8080/19864
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>