IMPALA-12123: Fix crash triggered by incomplete HDFS cache reads
When using HDFS caching, the HDFS cache may not have the full
buffer in memory, and it can return a buffer that is incomplete.
In this case, the code falls back to the ordinary read path.
However, the ScanRange cache_ structure is still set up, and
the code in ScanRange::ReadSubRanges() tries to use it. This
can crash, because the buffer is too short (and may have been
freed).

This changes the code to null out the cache_ data structure
when there is an incomplete read from the HDFS cache.

Testing:
 - Reproduced the crash stack manually by putting a Parquet
   file with a page index in HDFS cache and manually forcing
   it down the incomplete read codepath.
 - Modified the disk-io-mgr-test and CacheReaderTestStub to
   simulate the incomplete read case. The test will hit a
   DCHECK or crash without this fixup.

Change-Id: I51d8be6c03716badee81675447ed94ae6249b21b
Reviewed-on: http://gerrit.cloudera.org:8080/19869
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Zoltan Borok-Nagy <boroknagyz@cloudera.com>