IMPALA-12159: Support ORDER BY for collections of variable length types in select list
IMPALA-12019 implemented support for collections of fixed length types
in the sorting tuple. This change implements it for collections of
variable length types.

Note that the limitation that structs that contain any type of
collection are not allowed in the sorting tuple is still in place (see
IMPALA-12160).

Note that it was not and still is not allowed to sort by complex types,
this change only allows them to be present in the select list when
sortin by some other expression.

This change also allows collections of variable length types to be
non-passthrough children of UNION ALL nodes.

Testing:
 - Renamed the 'simple_arrays_big' table to 'arrays_big' and extended it
   with collections containing variable length types. This table is
   mainly used to test that spilling works during sorting.
 - Renamed
   test_sort.py::TestArraySort::{test_simple_arrays,
   test_simple_arrays_with_limit}
   to {test_array_sort,test_array_sort_with_limit}
 - Extended the tests run in test_queries.py::TestQueries::{test_sort,
   test_top_n,test_partitioned_top_n} with collections containing
   var-len types.
 - Added tests in sort-complex.test that assert that it is not allowed
   to sort by collections. For structs we already have such tests in
   struct-in-select-list.test.

Change-Id: Ic15b29393f260b572e11a8dbb9deeb8c02981852
Reviewed-on: http://gerrit.cloudera.org:8080/20108
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>