IMPALA-12228: Simulate the failure of an iceberg transaction.
The commit of an Iceberg transaction is done by the Iceberg catalog. In
the common case for Impala the Iceberg catalog is HiveCatalog, and the
actual commit is performed by HMS. This means the commit could fail
because of some activity outside of Impala. It is useful therefore to
be able to simulate what happens when an Iceberg commit fails.

Extend Java DebugAction to allow it to throw an exception. For now this
is limited to throwing unchecked exceptions, which is all that is
needed for this patch.

Add two DebugActions that can be used to throw Iceberg
CommitFailedExceptions at the point where the Iceberg transaction is
about to commit.

Add a new test that uses the new DebugActions to abort an insert and the
addition of a column.

Change-Id: Iafdacc3377a0a946ead9331ba6a63a1fbb11f0eb
Reviewed-on: http://gerrit.cloudera.org:8080/20306
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>