IMPALA-12158: Restore process name for daemons
IMPALA-10794 introduced names for threads, which messed up process names
too for catalogd and impalad.

Sets a pseudo-property sun.java.command that is used by jps to read the
Java process name. Use that property to replace the name of the "main"
thread (which, when updated, also updates the process name).

The process name is also used for renames in ThreadNameAnnotator so that
the original process name is always present. This ensures tools like
pgrep can find impalad and catalogd.

Change-Id: I7469b4f595acc1949286520a8e6086716bec0b1f
Reviewed-on: http://gerrit.cloudera.org:8080/19990
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Michael Smith <michael.smith@cloudera.com>