IMPALA-12193: Fix TSAN failures for DataCacheTest.SetReadOnly
The DataCacheTest.SetReadOnly currently fails in TSAN
builds, because there is a data race on the DataCache::readonly_
variable. The set of the variable happens without obtaining a
lock, and that is an intentional choice to prevent new readers
before waiting for existing readers to finish. This changes the
DataCache::readonly_ variable and the DataCache::CacheFile::readonly_
variable to be AtomicBools.

Testing:
 - DataCacheTest.SetReadOnly passes on a TSAN build.

Change-Id: Ibc507f09b9c093a9034d601d8e7d37976bd0433e
Reviewed-on: http://gerrit.cloudera.org:8080/20025
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>