IMPALA-12148: Bump timeouts for TestAsyncDDLTiming::test_ctas()
TestAsyncDDLTiming::test_ctas() has been failing intermittently
on ASAN test jobs and has been failing sporadically on all other
types of test jobs. The test is running in parallel with other
tests, so it is possible for the statement to take longer than
expected. This bumps the timeout to avoid this flakiness. The
actual test case is not testing that the statement is finishing
in a particular amount of time.

Testing:
 - Ran ASAN job

Change-Id: I4030afad3a00c4492bf7296bde7b77e86437111f
Reviewed-on: http://gerrit.cloudera.org:8080/20030
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>