IMPALA-12198: Create $RANGER_LOG_DIR before stopping in kill-ranger-server.sh
If the $RANGER_LOG_DIR directory doesn't exist, then Ranger's
ranger-admin-services.sh will fail to issue the stop command
because it tries and fails to write to that directory. Ranger's
script believes that it has issued the stop to Ranger, so it
waits for 30 seconds for Ranger to stop. When Ranger doesn't stop,
it kills Ranger. This is an unnecessary delay in stopping Ranger.
It is common if a developer runs bin/clean.sh after starting
Ranger.

This modifies kill-ranger-server.sh to create $RANGER_LOG_DIR
before running ranger-admin-services.sh with the stop command.
Since the directory exists, the stop command is successfully
issued and the script won't wait 30 seconds.

Testing:
 - Hand testing with starting Ranger, then running bin/clean.sh,
   then running kill-ranger-server.sh

Change-Id: I6ba5a90172affde3949e9f9a7618bde0dfa8c309
Reviewed-on: http://gerrit.cloudera.org:8080/20028
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>