IMPALA-12210: CatalogD should ignore validWriteId List if the table is external while loading a table
In external frontend mode, running queries on external tables shows an
exception in the catalogd.INFO log and catalogd forwards the
get_table_req() requests to HMS. For external tables, we don't need to
verify the validWriteIdList. So a transactional check on the table is
added in CatalogD before checking for compactionId on the table. By
doing so, when the table is loaded, catalogD returns the table instead
of forwarding the request to HMS.

Testing: Manually verified that exception is not seen in the logs.
Also, repeated execution of query, the request is not forwarded to HMS
and instead returned from catalogD's cache. Testing with TPCH queries
resulted in query performance improvement of ~ 0.04 to 0.07 secs.

Change-Id: I28a34185c788a39aa902b0cd1ed0ba0c8c376ff1
Reviewed-on: http://gerrit.cloudera.org:8080/20066
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Quanlong Huang <huangquanlong@gmail.com>