IMPALA-12179 (part 2): Use /etc/os-release for bootstrap_toolchain.py
This removes the usage of lsb_release in bootstrap_toolchain.py
in favor of reading /etc/os-release. /etc/os-release is available
on all distributions that we support. This combines the ID
and the major version component of VERSION_ID to produce the
distribution identifier. A few example:
Ubuntu 16.04: ID=ubuntu, VERSION_ID=16.04 => ubuntu16
Rocky 8.5: ID=rocky, VERSION_ID=8.5 => rocky8
SLES 15.1: ID=sles, VERSION_ID=15.1 => sles15

As cleanup, this removes old distributions that we no longer
support (e.g. Redhat 5/6, Debian, Sles 11, etc). It also removes
the unused CDH component of the OS_MAPPING.

The values used in OS_MAPPING are based on the database of
/etc/os-release files available at https://github.com/chef/os_release

Testing:
 - Ran the logic against the /etc/os-release files for Redhat/Centos 7,
   Redhat/Centos/Rocky/Almalinux 8, Ubuntu, and SLES 12/15.

Change-Id: Ida3ffb8525c5b750ddbf9fd3ed5d0782fac9cdd0
Reviewed-on: http://gerrit.cloudera.org:8080/20070
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>