IMPALA-11949: Make backend AVRO C++ library available
The C++ library has been added earlier to the toolchain.

This is a small commit to allow downloading the C++ Avro library with
the toolchain and starting a build with it. Its aim is to help
transitioning from C to C++ library, once the transition is complete
this environment variable can be deprecated and removed.

Testing:
 - Manualy tested switching and downloading the package

Change-Id: I397341db1ea7195db3649f5220a3ffe8a3919681
Reviewed-on: http://gerrit.cloudera.org:8080/20097
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>